+++ node/src/Consumer.ts
@@ -899,7 +899,7 @@ export class Consumer<
 					}
 
 					default: {
-						logger.error('ignoring unknown event "%s"', event);
+						logger.error(`ignoring unknown event "${event}"`);
 					}
 				}
 			}