+++ node/src/DirectTransport.ts
@@ -276,7 +276,7 @@ export class DirectTransport<
 					}
 
 					default: {
-						logger.error('ignoring unknown event "%s"', event);
+						logger.error(`ignoring unknown event "${event}"`);
 					}
 				}
 			}