+++ node/src/PipeTransport.ts
@@ -450,7 +450,7 @@ export class PipeTransport<
 					}
 
 					default: {
-						logger.error('ignoring unknown event "%s"', event);
+						logger.error(`ignoring unknown event "${event}"`);
 					}
 				}
 			}