+++ node/src/PlainTransport.ts
@@ -451,7 +451,7 @@ export class PlainTransport<
 					}
 
 					default: {
-						logger.error('ignoring unknown event "%s"', event);
+						logger.error(`ignoring unknown event "${event}"`);
 					}
 				}
 			}