+++ node/src/WebRtcTransport.ts
@@ -644,7 +644,7 @@ export class WebRtcTransport<
 					}
 
 					default: {
-						logger.error('ignoring unknown event "%s"', event);
+						logger.error(`ignoring unknown event "${event}"`);
 					}
 				}
 			}