+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `Worker`: Fix `io_uring` support detection ([PR #1445](https://github.com/versatica/mediasoup/pull/1445)).
 - Mitigate libsrtp wraparound with loss decryption failure ([PR #1438](https://github.com/versatica/mediasoup/pull/1438)).
+- Node: New `setLogEventListeners()` utility to get log events ([PR #XXXX](https://github.com/versatica/mediasoup/pull/XXXX)).
 
 ### 3.14.11